create_ffi_helper_set
